introduce ilk.dust: minimum debt requirement
It is sometimes useful to require that an ilk has a minimum level of
debt per urn, other than zero. A concrete example is to prevent griefing
of the system by the creation of many dusty urns.

Note that it is still possible for dusty urns to exist: for example,
a partial `bite` could leave tab < dust. In this state any subsequent
operation by the owner must either increase their debt over the
threshold or clear it entirely, resulting in tab > dust or tab == 0.