bite: partial liquidations
Each bite now starts an auction to cover a fixed amount of debt.

This combines the previous `bite` and `flip` methods into one. Prior to
this change liquidations were of the full CDP, which required additional
state tracking of the liquidated position. With partial liquidations,
large CDPs could potentially become safe again if the price recovers
before the position is fully liquidated.